Added uid-common files
