+++ components/uid-common/README.md
@@ -1,0 +1,16 @@
+# Common library for UID web components
+
+This library contents common elements used in UID web components.
+
+
+## Installation
+
+```bash
+npm i uid-common
+```
+
+## Usage
+
+Import the required class or file from an UID web component.
+
+
