+++ components/uid-common/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@bonitasoft/uid-common",
+  "description": "Common library for UID web components",
+  "license": "GPLv2",
+  "version": "1.0.0",
+  "scripts": {
+    "build": "tsc",
+    "test": "tsc",
+    "prepublish": "tsc"
+  },
+  "dependencies": {
+    "lit": "^2.0.0-rc.2"
+  },
+  "devDependencies": {
+    "tslib": "^2.3.1",
+    "typescript": "^4.3.5"
+  }
+}
